Try to make test_storage import on Python 3.
